+++ vector/src/main/java/im/vector/view/PasswordStrengthBar.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package im.vector.ui
+package im.vector.view
 
 import android.content.Context
 import android.util.AttributeSet
