+++ vector/src/main/res/layout/keys_backup_setup_step2_fragment.xml
@@ -62,7 +62,7 @@
             app:layout_constraintStart_toEndOf="@id/keys_backup_passphrase_enter_til"
             app:layout_constraintTop_toTopOf="@+id/keys_backup_passphrase_enter_til" />
 
-        <im.vector.ui.PasswordStrengthBar
+        <im.vector.view.PasswordStrengthBar
             android:id="@+id/keys_backup_passphrase_security_progress"
             android:layout_width="0dp"
             android:layout_height="4dp"
